% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpkg.R
\docType{data}
\name{survey}
\alias{build}
\alias{survey}
\title{This is a Mason template for a generic R package}
\format{\preformatted{List of 18
 $ name           :List of 2
  ..$ expr: language input("Name:", default = basename(getwd()))
  ..$ env :<environment: namespace:mason.rpkg> 
  ..- attr(*, "class")= chr "lazy"
 $ title          :List of 2
  ..$ expr: language input("Title:", default = answers$name)
  ..$ env :<environment: namespace:mason.rpkg> 
  ..- attr(*, "class")= chr "lazy"
 $ version        :List of 2
  ..$ expr: language input("Version:", default = "1.0.0")
  ..$ env :<environment: namespace:mason.rpkg> 
  ..- attr(*, "class")= chr "lazy"
 $ author         :List of 2
  ..$ expr: language input("Author:", default = default_author())
  ..$ env :<environment: namespace:mason.rpkg> 
  ..- attr(*, "class")= chr "lazy"
 $ maintainer     :List of 2
  ..$ expr: language input("Maintainer:", default = default_maintainer(answers))
  ..$ env :<environment: namespace:mason.rpkg> 
  ..- attr(*, "class")= chr "lazy"
 $ description    :List of 2
  ..$ expr: language input("Description:", default = answers$title)
  ..$ env :<environment: namespace:mason.rpkg> 
  ..- attr(*, "class")= chr "lazy"
 $ license        :List of 2
  ..$ expr: language choose("License:", licenses, default = "MIT")
  ..$ env :<environment: namespace:mason.rpkg> 
  ..- attr(*, "class")= chr "lazy"
 $ url            :List of 2
  ..$ expr: language input("URL:", default = default_url(answers))
  ..$ env :<environment: namespace:mason.rpkg> 
  ..- attr(*, "class")= chr "lazy"
 $ bugreports     :List of 2
  ..$ expr: language input("BugReports:", default = default_bugreports(answers))
  ..$ env :<environment: namespace:mason.rpkg> 
  ..- attr(*, "class")= chr "lazy"
 $ testing        :List of 2
  ..$ expr: language choose("Testing framework:", choices = c("testthat", "none"), default = 1)
  ..$ env :<environment: namespace:mason.rpkg> 
  ..- attr(*, "class")= chr "lazy"
 $ readme         :List of 2
  ..$ expr: language confirm("README.md file:", default = TRUE)
  ..$ env :<environment: namespace:mason.rpkg> 
  ..- attr(*, "class")= chr "lazy"
 $ readme_rmd     :List of 2
  ..$ expr: language confirm("README.Rmd file:", default = TRUE, when = function(a) a$readme)
  ..$ env :<environment: namespace:mason.rpkg> 
  ..- attr(*, "class")= chr "lazy"
 $ news           :List of 2
  ..$ expr: language confirm("NEWS.md file:", default = TRUE)
  ..$ env :<environment: namespace:mason.rpkg> 
  ..- attr(*, "class")= chr "lazy"
 $ cis            :List of 2
  ..$ expr: language checkbox("Continuous integration:", choices = cis, default = c("Travis",      "Appveyor"))
  ..$ env :<environment: namespace:mason.rpkg> 
  ..- attr(*, "class")= chr "lazy"
 $ create_git_repo:List of 2
  ..$ expr: language confirm("Create git repo?", default = TRUE)
  ..$ env :<environment: namespace:mason.rpkg> 
  ..- attr(*, "class")= chr "lazy"
 $ create_gh_repo :List of 2
  ..$ expr: language confirm("Create repo on GitHub?", default = TRUE, when = function(a) a$create_git_repo)
  ..$ env :<environment: namespace:mason.rpkg> 
  ..- attr(*, "class")= chr "lazy"
 $ gh_username    :List of 2
  ..$ expr: language input("GitHub username:", default = whoami::gh_username(), when = function(a) a$create_git_repo &&      a$create_gh_repo)
  ..$ env :<environment: namespace:mason.rpkg> 
  ..- attr(*, "class")= chr "lazy"
 $ push_to_github :List of 2
  ..$ expr: language confirm("Push initial version to GitHub?", default = FALSE, when = function(a) a$create_gh_repo)
  ..$ env :<environment: namespace:mason.rpkg> 
  ..- attr(*, "class")= chr "lazy"
 - attr(*, "class")= chr "ask_questions"
}}
\usage{
survey

build(answers)
}
\arguments{
\item{answers}{The answers the builder operates on.}
}
\description{
The template somewhat leans towards GitHub, actually,
so it is best for packages developed at GitHub.
}
\keyword{datasets}

