% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpkg.R
\docType{data}
\name{survey}
\alias{build}
\alias{survey}
\title{This is a Mason template for a generic R package}
\format{An object of class \code{ask_questions} of length 25.}
\usage{
survey

build(answers)
}
\arguments{
\item{answers}{The answers the builder operates on.}
}
\description{
The template somewhat leans towards GitHub, actually,
so it is best for packages developed at GitHub.
}
\keyword{datasets}

